Have dedicated inserters for different entity types
This finishes https://github.com/wmde/WikibaseQueryEngine/issues/40 for the
code internal to the store. Now all that is needed is updating the public
interface

The same mechanism is also present for removal, though
just has a single trivial implementation for all entity types
so far. This can easily be changed once the store needs it.

The dedicated class for updating was removed, rather than
made more elaborate, since the current thing done is trivial
and can be done for all entity types.